hacker-news-analyzer/pyproject.toml
[project]
name = "hacker-news-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain~=1.0",
    "langchain-openai~=1.0",
    "python-dotenv!=1.1",
]
